dividing button widget and adding stateful widget
